Refactor RoomItem
